#230 - add fallback bridge component to log events without portal
