+++ ipsframework/bridges/__init__.py
@@ -1,0 +1,6 @@
+"""Bridges are components which handle supplementary tasks in the IPS Framework. Applications should not need to import these classes directly.
+
+There are two bridges available:
+- `BasicBridge`, which provides simple system logging
+- `PortalBridge`, which provides all functionality of `BasicBridge` and additionally allows interfacing with a remote IPS Portal
+"""
